requirements.txt
# Core trading dependencies
alpaca-py>=0.8.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
yfinance>=0.2.0
pytz==2023.3
requests==2.31.0
websocket-client==1.6.1
backtrader==1.9.78.123
scikit-learn==1.3.0
pytest==7.4.0
ibapi==9.81.1.post1
ib_async>=0.9.86

# Multi-agent AI dependencies
typing-extensions
langchain-openai
langchain-experimental
langchain_anthropic
langchain-google-genai
langchain-deepseek
langgraph
chromadb
setuptools

# Data sources and web scraping
praw
feedparser
stockstats
eodhd
akshare
tushare
finnhub-python
parsel
tqdm

# Storage and caching
redis

# UI and CLI
chainlit
rich
questionary
# typer


